+++ scientific/Changelog.md
@@ -7,6 +7,7 @@
 * Reorder code
 * Use actual end for static Ptr/debug
 * Use NonNull instead of a raw ptr
+* Add TryFrom for u/isize, improve From<u/isize>
 
 ## 0.1.2 -- 2021-07-08